fix(llma): do not pass posthog params
