+++ .changeset/dirty-trains-slide.md
@@ -1,0 +1,5 @@
+---
+'@posthog/ai': patch
+---
+
+Ensure Posthog parameters are not passed to any provider
